Refactor folding and barrier
